+++ src/common/strategies/drawrectanglestrategy.cpp
@@ -8,8 +8,8 @@ void DrawRectangleStrategy::drawRequest(QPainter *painter)
 void DrawRectangleStrategy::mouseTouch(const PaintTypes::MouseState &state, bool paintStarted)
 {
     if (!paintStarted) {
-        m_bounds.setStart({static_cast<qreal>(state.m_x), static_cast<qreal>(state.m_y)});
+        m_bounds.setStart({state.m_x, state.m_y});
     } else {
-        m_bounds.setEnd({static_cast<qreal>(state.m_x), static_cast<qreal>(state.m_y)});
+        m_bounds.setEnd({state.m_x, state.m_y});
     }
 }
